# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.NftCollection:
    properties:
      chain:
        $ref: '#/definitions/model.Chain'
      chain_id:
        type: integer
      collection_address:
        type: string
      contract_scan:
        type: string
      created_time:
        type: string
      description:
        type: string
      discord:
        type: string
      erc_format:
        type: string
      id:
        type: integer
      image:
        type: string
      is_notify_synced:
        type: boolean
      is_rarity_calculated:
        type: boolean
      last_updated_time:
        type: string
      marketplace:
        items:
          $ref: '#/definitions/model.NftListingMarketplace'
        type: array
      marketplace_collections:
        items:
          $ref: '#/definitions/model.NftMarketplaceCollection'
        type: array
      name:
        type: string
      owners:
        type: integer
      stats:
        $ref: '#/definitions/model.ViewNftCollectionStats'
      supply:
        type: integer
      symbol:
        type: string
      twitter:
        type: string
      website:
        type: string
    type: object
...
  model.NftMetadataAttributesIcon:
    properties:
      created_at:
        type: string
      discord_icon:
        type: string
      id:
        type: integer
      trait_type:
        type: string
      unicode_icon:
        type: string
      updated_at:
        type: string
    type: object
...
  model.NftToken:
    properties:
      amount:
        type: string
      attributes:
        description: disable write
        items:
          $ref: '#/definitions/model.NftTokenAttribute'
        type: array
      collection_address:
        type: string
      created_time:
        type: string
      description:
        type: string
      image:
        type: string
      image_cdn:
        type: string
      image_content_type:
        type: string
      is_self_hosted:
        type: boolean
      last_updated_time:
        type: string
      listing:
        $ref: '#/definitions/model.NftListing'
        description: Listing    *NftListing         `json:"listing" gorm:"foreignKey:TokenId,ContractAddress;references:TokenId,CollectionAddress;<-:false"`
          // disable write
      listing_marketplace:
        items:
          $ref: '#/definitions/model.NftListingMarketplace'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/model.NftOwner'
      rarity:
        $ref: '#/definitions/model.NftTokenRarity'
        description: Attributes []NftTokenAttribute `json:"attributes" gorm:"foreignkey:TokenId,CollectionAddress;references:TokenId,CollectionAddress;<-:false"`
          // disable write
      rarity_rank:
        type: integer
      rarity_score:
        type: string
      rarity_tier:
        type: string
      thumbnail_cdn:
        type: string
      token_id:
        type: string
    type: object
...
  response.GetNftCollectionMetadataResponse:
    properties:
      data:
        $ref: '#/definitions/model.NftCollection'
    type: object
  response.GetNftCollectionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NftCollection'
        type: array
      page:
        description: page index
        type: integer
      size:
        description: page size
        type: integer
      total:
        type: integer
    type: object
  response.GetNftTokenActivitiesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NftListing'
        type: array
      page:
        description: page index
        type: integer
      size:
        description: page size
        type: integer
      total:
        type: integer
    type: object
  response.GetNftTokenTransactionHistory:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NftTxHistory'
        type: array
    type: object
  response.GetNftTokensResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NftToken'
        type: array
      page:
        description: page index
        type: integer
      size:
        description: page size
        type: integer
      total:
        type: integer
    type: object
...
  response.NftCollectionTickersResponse:
    properties:
      data:
        $ref: '#/definitions/response.NftCollectionTickersData'
    type: object
  response.NftMetadataAttributesIconResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NftMetadataAttributesIcon'
        type: array
      page:
        description: page index
        type: integer
      size:
        description: page size
        type: integer
      total:
        type: integer
    type: object
...
  response.NftTokenResponse:
    properties:
      data:
        $ref: '#/definitions/response.NftTokenData'
    type: object
...
  response.NftTokenTickersDataResponse:
    properties:
      data:
        $ref: '#/definitions/response.NftTokenTickersData'
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /{wallet_address}/nft:
    get:
      consumes:
      - application/json
      description: Get nft tokens by owner address
      parameters:
      - description: wallet address
        in: path
        name: wallet_address
        required: true
        type: string
      - description: chain ID
        in: query
        name: chain_id
        type: integer
      - in: query
        items:
          type: string
        name: collection_addresses
        type: array
      - description: page index
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: size
        type: integer
      - description: sort by
        enum:
        - collection_address
        - token_id
        - name
        - amount
        - rarity_rank
        - rarity_score
        - rarity_tier
        - -collection_address
        - -token_id
        - -name
        - -amount
        - -rarity_rank
        - -rarity_score
        - -rarity_tier
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetNftTokensResponse'
      summary: Get nft tokens by owner address
      tags:
      - Wallet
...
  /nft:
    get:
      consumes:
      - application/json
      description: Get nft collections
      parameters:
      - description: collection address
        in: query
        name: address
        type: string
      - description: marketplace
        enum:
        - paintswap
        - nftkey
        - opensea
        - looksrare
        in: query
        name: marketplace
        type: string
      - description: search collection name
        in: query
        name: name
        type: string
      - description: page index
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: size
        type: integer
      - enum:
        - one_day_volume
        - -one_day_volume
        - seven_day_volume
        - -seven_day_volume
        - thirty_day_volume
        - -thirty_day_volume
        - all_time_volume
        - -all_time_volume
        - floor_price
        - -floor_price
        - supply
        - -supply
        - name
        - -name
        - owners
        - -owners
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetNftCollectionsResponse'
      summary: Get nft collections
      tags:
      - NFT
  /nft/{collection_address}:
    get:
      consumes:
      - application/json
      description: Get tokens of a nft collection
      parameters:
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      - collectionFormat: multi
        description: marketplaces
        in: query
        items:
          enum:
          - paintswap
          - nftkey
          - opensea
          - looksrare
          type: string
        name: marketplaces
        type: array
      - collectionFormat: multi
        description: token traits/attributes
        example: mask=Television Head,Zorro Mask
        in: query
        items:
          type: string
        name: traits
        type: array
      - collectionFormat: multi
        description: traits count
        example: 11
        in: query
        items:
          type: integer
        name: traits_count
        type: array
      - description: payment token/currency
        enum:
        - eth
        - ftm
        in: query
        name: currency
        type: string
      - description: listing status
        enum:
        - listed
        - sold
        - cancelled
        in: query
        name: listing_status
        type: string
      - description: listing type
        enum:
        - buy_now
        - auction
        in: query
        name: listing_type
        type: string
      - description: token name
        in: query
        name: name
        type: string
      - description: page index
        in: query
        name: page
        type: integer
      - description: listing price max
        example: 10
        in: query
        name: price_max
        type: number
      - description: listing price min
        example: 0
        in: query
        name: price_min
        type: number
      - description: rarity rank max
        in: query
        name: rarity_rank_max
        type: integer
      - description: rarity rank min
        in: query
        name: rarity_rank_min
        type: integer
      - description: page size
        in: query
        name: size
        type: integer
      - description: sort by
        enum:
        - name
        - -name
        - token_id
        - -token_id
        - amount
        - -amount
        - rarity_rank
        - -rarity_rank
        - rarity_score
        - -rarity_score
        - rarity_tier
        - -rarity_tier
        in: query
        name: sort
        type: string
      - description: token ID
        in: query
        name: token_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetNftTokensResponse'
      summary: Get tokens of a nft collection
      tags:
      - NFT
  /nft/{collection_address}/{token_id}:
    get:
      consumes:
      - application/json
      description: Get nft details
      parameters:
      - description: token Id
        in: path
        name: token_id
        required: true
        type: integer
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NftTokenResponse'
      summary: Get nft by address
      tags:
      - NFT
  /nft/{collection_address}/{token_id}/activity:
    get:
      consumes:
      - application/json
      description: Get nft token activities
      parameters:
      - description: token Id
        in: path
        name: token_id
        required: true
        type: integer
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      - description: marketplace name
        in: query
        name: marketplace
        type: string
      - description: page index
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetNftTokenActivitiesResponse'
      summary: Get nft token activities
      tags:
      - NFT
  /nft/{collection_address}/{token_id}/transaction-history:
    get:
      consumes:
      - application/json
      description: Get nft token tx history
      parameters:
      - description: token Id
        in: path
        name: token_id
        required: true
        type: integer
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetNftTokenTransactionHistory'
      summary: Get nft token tx history
      tags:
      - NFT
  /nft/{collection_address}/metadata:
    get:
      consumes:
      - application/json
      description: Get nft collection metadata
      parameters:
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetNftCollectionMetadataResponse'
      summary: Get nft collection metadata
      tags:
      - NFT
  /nft/metadata/attributes-icon:
    get:
      consumes:
      - application/json
      description: Get nft trait icons
      parameters:
      - in: query
        items:
          type: string
        name: trait_type
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NftMetadataAttributesIconResponse'
      summary: Get nft trait icons
      tags:
      - NFT
...
  /nft/ticker/{collection_address}:
    get:
      consumes:
      - application/json
      description: Get floor prices snapshot by nft address (top 7d / 1m / ...)
      parameters:
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      - description: From date in UNIX millisecond (eg. 1392577232000)
        in: query
        name: from
        required: true
        type: integer
      - description: To date in UNIX millisecond (eg. 1422577232000)
        in: query
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NftCollectionTickersResponse'
      summary: Get floor prices snapshot by nft address (top 7d / 1m / ...)
      tags:
      - NFT
  /nft/ticker/{collection_address}/{token_id}:
    get:
      consumes:
      - application/json
      description: Get nft token ticker
      parameters:
      - description: collection address
        in: path
        name: collection_address
        required: true
        type: string
      - description: token_id
        in: path
        name: token_id
        required: true
        type: string
      - description: From date in UNIX millisecond (eg. 1392577232000)
        in: query
        name: from
        required: true
        type: integer
      - description: To date in UNIX millisecond (eg. 1422577232000)
        in: query
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NftTokenTickersDataResponse'
      summary: Get nft token ticker
      tags:
      - NFT
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"